# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brightloom API",
    "version": "1.0.0",
    "summary": "Tasks, projects and billing for small teams.",
    "description": "The HTTP API behind the Brightloom web app. Every endpoint apart from the login call and the billing webhook needs the session cookie that POST /api/v1/auth/login sets.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Brightloom engineering",
      "email": "engineering@brightloom.test"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Local development server"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "listTasks",
        "summary": "List tasks with optional filters",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "description": "Only tasks in this project",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "description": "Only tasks assigned to this user",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Only tasks in this status",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "done"
              ]
            }
          },
          {
            "name": "due",
            "in": "query",
            "description": "A due date window relative to today",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "upcoming",
                "overdue"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many tasks to return",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "How many tasks to skip",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "operationId": "createTask",
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/tasks/completed/stats": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "completedTaskStats",
        "summary": "Count completed tasks per day for the last fourteen days",
        "responses": {
          "200": {
            "description": "One entry per day, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletedStatsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Task id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": [
          "Tasks"
        ],
        "operationId": "getTask",
        "summary": "Read one task with its comments",
        "responses": {
          "200": {
            "description": "The task and its comments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "operationId": "updateTask",
        "summary": "Edit, complete or reschedule a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "operationId": "deleteTask",
        "summary": "Delete a task and its comments",
        "responses": {
          "200": {
            "description": "The task was deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/tasks/{id}/comments": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Task id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "post": {
        "tags": [
          "Tasks"
        ],
        "operationId": "createComment",
        "summary": "Add a comment to a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created comment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bl_session",
        "description": "The session cookie set by POST /api/v1/auth/login."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was not valid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No valid session cookie was sent",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No record matches that id",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "not_found"
                ]
              },
              "message": {
                "type": "string",
                "examples": [
                  "That task does not exist"
                ]
              }
            }
          }
        }
      },
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "project_id",
          "title",
          "description",
          "status",
          "priority",
          "due_date",
          "assignee_id",
          "created_at",
          "completed_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "org_id": {
            "type": "integer"
          },
          "project_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_color": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done"
            ]
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "examples": [
              "2026-01-31"
            ]
          },
          "assignee_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "assignee_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "Comment": {
        "type": "object",
        "required": [
          "id",
          "task_id",
          "author_id",
          "author_name",
          "body",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "task_id": {
            "type": "integer"
          },
          "author_id": {
            "type": "integer"
          },
          "author_name": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TaskCreateRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "project_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "examples": [
              "2026-01-31"
            ]
          },
          "assignee_id": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "TaskUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "done"
            ]
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4
          },
          "due_date": {
            "type": [
              "string",
              "null"
            ],
            "examples": [
              "2026-01-31"
            ]
          },
          "assignee_id": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "required": [
          "task"
        ],
        "properties": {
          "task": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
      "TaskDetailResponse": {
        "type": "object",
        "required": [
          "task",
          "comments"
        ],
        "properties": {
          "task": {
            "$ref": "#/components/schemas/Task"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        }
      },
      "TaskListResponse": {
        "type": "object",
        "required": [
          "tasks",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
      "CommentCreateRequest": {
        "type": "object",
        "required": [
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
      "CommentResponse": {
        "type": "object",
        "required": [
          "comment"
        ],
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/Comment"
          }
        }
      },
      "CompletedStatsResponse": {
        "type": "object",
        "required": [
          "days"
        ],
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "date",
                "count"
              ],
              "properties": {
                "date": {
                  "type": "string",
                  "examples": [
                    "2026-01-31"
                  ]
                },
                "count": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
...
    }
  }
}